+++ src/components/fields/Field_SingleRelationship.vue
@@ -98,7 +98,7 @@
                       <q-icon name="mdi-format-color-fill" />
                     </q-item-section>
                   </q-item>
-                  <q-separator />
+                  <q-separator dark />
                     <q-item clickable @click="openExistingInput(fixGetCorrectDocument(localInput))">
                     <q-item-section>Open document</q-item-section>
                     <q-item-section avatar>
@@ -214,7 +214,7 @@
                       <q-icon name="mdi-format-color-fill" />
                     </q-item-section>
                   </q-item>
-                  <q-separator />
+                  <q-separator dark />
                     <q-item clickable @click="openExistingInput(fixGetCorrectDocument(scope.opt))">
                     <q-item-section>Open document</q-item-section>
                     <q-item-section avatar>
@@ -315,7 +315,7 @@
                       <q-icon name="mdi-format-color-fill" />
                     </q-item-section>
                   </q-item>
-                  <q-separator />
+                  <q-separator dark />
                     <q-item clickable @click="openExistingInput(opt)">
                     <q-item-section>Open document</q-item-section>
                     <q-item-section avatar>